Moved test_logsumexp to tests/tensor/test_special.py
